requirements/common.txt
boto3==1.26.118
boto3==1.26.118
celery[sqs]>=5.2.7, <6
coverage==7.2.7
dj-rest_auth==4.0.1
django-anymail==9.1
django-cors-headers==4.2.0
django-filter==23.2
django-simple-history==3.3.0
djangorestframework>=3.13.1, <4
djangorestframework-simplejwt>=5.2.2,<5.3.0
drf-yasg==1.21.5, <2
gunicorn==20.1.0
jinja2>=3.1.2, <4
pre-commit==3.3.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.30.0
sib-api-v3-sdk==7.6.0
tenacity>=8.0.0, <9
tox==4.6.4
uvicorn==0.20.0
pynamodb==6.0.0
